requirements.txt
asgiref==3.4.1
astroid==2.9.2
colorama==0.4.4
Django==4.0.1
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
djangorestframework-simplejwt
mysqlclient==2.1.0
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
wrapt==1.13.3
